# webapp/swagger.yaml
basePath: /api/v1
consumes:
    - application/json
definitions:
    APIErrorResponse:
        properties:
            details:
                type: string
                x-go-name: Details
            error:
                type: string
                x-go-name: Error
        type: object
        x-go-package: github.com/cloudbase/garm/apiserver/params
...
    CreateGiteaCredentialsParams:
        properties:
            app:
                $ref: '#/definitions/GithubApp'
            auth_type:
                $ref: '#/definitions/ForgeAuthType'
            description:
                type: string
                x-go-name: Description
            endpoint:
                type: string
                x-go-name: Endpoint
            name:
                type: string
                x-go-name: Name
            pat:
                $ref: '#/definitions/GithubPAT'
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    CreateGithubCredentialsParams:
        properties:
            app:
                $ref: '#/definitions/GithubApp'
            auth_type:
                $ref: '#/definitions/ForgeAuthType'
            description:
                type: string
                x-go-name: Description
            endpoint:
                type: string
                x-go-name: Endpoint
            name:
                type: string
                x-go-name: Name
            pat:
                $ref: '#/definitions/GithubPAT'
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    Credentials:
        description: used by swagger client generated code
        items:
            $ref: '#/definitions/ForgeCredentials'
        type: array
        x-go-package: github.com/cloudbase/garm/params
    EndpointType:
        type: string
        x-go-package: github.com/cloudbase/garm/params
    Enterprise:
        properties:
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            credentials:
                $ref: '#/definitions/ForgeCredentials'
            credentials_id:
                format: uint64
                type: integer
                x-go-name: CredentialsID
            credentials_name:
                description: |-
                    CredentialName is the name of the credentials associated with the enterprise.
                    This field is now deprecated. Use CredentialsID instead. This field will be
                    removed in v0.2.0.
                type: string
                x-go-name: CredentialsName
            endpoint:
                $ref: '#/definitions/ForgeEndpoint'
            events:
                items:
                    $ref: '#/definitions/EntityEvent'
                type: array
                x-go-name: Events
            id:
                type: string
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            pool:
                items:
                    $ref: '#/definitions/Pool'
                type: array
                x-go-name: Pools
            pool_balancing_type:
                $ref: '#/definitions/PoolBalancerType'
            pool_manager_status:
                $ref: '#/definitions/PoolManagerStatus'
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    EntityEvent:
        properties:
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            event_level:
                $ref: '#/definitions/EventLevel'
            event_type:
                $ref: '#/definitions/EventType'
            id:
                format: uint64
                type: integer
                x-go-name: ID
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-package: github.com/cloudbase/garm/params
    EventLevel:
        type: string
        x-go-package: github.com/cloudbase/garm/params
    EventType:
        type: string
        x-go-package: github.com/cloudbase/garm/params
    ForgeAuthType:
        type: string
        x-go-package: github.com/cloudbase/garm/params
    ForgeCredentials:
        properties:
            api_base_url:
                type: string
                x-go-name: APIBaseURL
            auth-type:
                $ref: '#/definitions/ForgeAuthType'
            base_url:
                type: string
                x-go-name: BaseURL
            ca_bundle:
                items:
                    format: uint8
                    type: integer
                type: array
                x-go-name: CABundle
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            description:
                type: string
                x-go-name: Description
            endpoint:
                $ref: '#/definitions/ForgeEndpoint'
            enterprises:
                items:
                    $ref: '#/definitions/Enterprise'
                type: array
                x-go-name: Enterprises
            forge_type:
                $ref: '#/definitions/EndpointType'
            id:
                format: uint64
                type: integer
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            organizations:
                items:
                    $ref: '#/definitions/Organization'
                type: array
                x-go-name: Organizations
            rate_limit:
                $ref: '#/definitions/GithubRateLimit'
            repositories:
                items:
                    $ref: '#/definitions/Repository'
                type: array
                x-go-name: Repositories
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
            upload_base_url:
                type: string
                x-go-name: UploadBaseURL
        type: object
        x-go-package: github.com/cloudbase/garm/params
    ForgeEndpoint:
        properties:
            api_base_url:
                type: string
                x-go-name: APIBaseURL
            base_url:
                type: string
                x-go-name: BaseURL
            ca_cert_bundle:
                items:
                    format: uint8
                    type: integer
                type: array
                x-go-name: CACertBundle
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            description:
                type: string
                x-go-name: Description
            endpoint_type:
                $ref: '#/definitions/EndpointType'
            name:
                type: string
                x-go-name: Name
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
            upload_base_url:
                type: string
                x-go-name: UploadBaseURL
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    GithubApp:
        properties:
            app_id:
                format: int64
                type: integer
                x-go-name: AppID
            installation_id:
                format: int64
                type: integer
                x-go-name: InstallationID
            private_key_bytes:
                items:
                    format: uint8
                    type: integer
                type: array
                x-go-name: PrivateKeyBytes
        type: object
        x-go-package: github.com/cloudbase/garm/params
    GithubPAT:
        properties:
            oauth2_token:
                type: string
                x-go-name: OAuth2Token
        type: object
        x-go-package: github.com/cloudbase/garm/params
    GithubRateLimit:
        properties:
            limit:
                format: int64
                type: integer
                x-go-name: Limit
            remaining:
                format: int64
                type: integer
                x-go-name: Remaining
            reset:
                format: int64
                type: integer
                x-go-name: Reset
            used:
                format: int64
                type: integer
                x-go-name: Used
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    OSArch:
        type: string
        x-go-package: github.com/cloudbase/garm-provider-common/params
    OSType:
        type: string
        x-go-package: github.com/cloudbase/garm-provider-common/params
    Organization:
        properties:
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            credentials:
                $ref: '#/definitions/ForgeCredentials'
            credentials_id:
                format: uint64
                type: integer
                x-go-name: CredentialsID
            credentials_name:
                description: |-
                    CredentialName is the name of the credentials associated with the enterprise.
                    This field is now deprecated. Use CredentialsID instead. This field will be
                    removed in v0.2.0.
                type: string
                x-go-name: CredentialsName
            endpoint:
                $ref: '#/definitions/ForgeEndpoint'
            events:
                items:
                    $ref: '#/definitions/EntityEvent'
                type: array
                x-go-name: Events
            id:
                type: string
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            pool:
                items:
                    $ref: '#/definitions/Pool'
                type: array
                x-go-name: Pools
            pool_balancing_type:
                $ref: '#/definitions/PoolBalancerType'
            pool_manager_status:
                $ref: '#/definitions/PoolManagerStatus'
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    Pool:
        properties:
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            enabled:
                type: boolean
                x-go-name: Enabled
            endpoint:
                $ref: '#/definitions/ForgeEndpoint'
            enterprise_id:
                type: string
                x-go-name: EnterpriseID
            enterprise_name:
                type: string
                x-go-name: EnterpriseName
            extra_specs:
                description: |-
                    ExtraSpecs is an opaque raw json that gets sent to the provider
                    as part of the bootstrap params for instances. It can contain
                    any kind of data needed by providers. The contents of this field means
                    nothing to garm itself. We don't act on the information in this field at
                    all. We only validate that it's a proper json.
                type: object
                x-go-name: ExtraSpecs
            flavor:
                type: string
                x-go-name: Flavor
            github-runner-group:
                description: |-
                    GithubRunnerGroup is the github runner group in which the runners will be added.
                    The runner group must be created by someone with access to the enterprise.
                type: string
                x-go-name: GitHubRunnerGroup
            id:
                type: string
                x-go-name: ID
            image:
                type: string
                x-go-name: Image
            instances:
                items:
                    $ref: '#/definitions/Instance'
                type: array
                x-go-name: Instances
            max_runners:
                format: uint64
                type: integer
                x-go-name: MaxRunners
            min_idle_runners:
                format: uint64
                type: integer
                x-go-name: MinIdleRunners
            org_id:
                type: string
                x-go-name: OrgID
            org_name:
                type: string
                x-go-name: OrgName
            os_arch:
                $ref: '#/definitions/OSArch'
            os_type:
                $ref: '#/definitions/OSType'
            priority:
                description: |-
                    Priority is the priority of the pool. The higher the number, the higher the priority.
                    When fetching matching pools for a set of tags, the result will be sorted in descending
                    order of priority.
                format: uint64
                type: integer
                x-go-name: Priority
            provider_name:
                type: string
                x-go-name: ProviderName
            repo_id:
                type: string
                x-go-name: RepoID
            repo_name:
                type: string
                x-go-name: RepoName
            runner_bootstrap_timeout:
                format: uint64
                type: integer
                x-go-name: RunnerBootstrapTimeout
            runner_prefix:
                type: string
                x-go-name: Prefix
            tags:
                items:
                    $ref: '#/definitions/Tag'
                type: array
                x-go-name: Tags
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
        type: object
        x-go-package: github.com/cloudbase/garm/params
    PoolBalancerType:
        type: string
        x-go-package: github.com/cloudbase/garm/params
    PoolManagerStatus:
        properties:
            failure_reason:
                type: string
                x-go-name: FailureReason
            running:
                type: boolean
                x-go-name: IsRunning
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    Repository:
        properties:
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            credentials:
                $ref: '#/definitions/ForgeCredentials'
            credentials_id:
                format: uint64
                type: integer
                x-go-name: CredentialsID
            credentials_name:
                description: |-
                    CredentialName is the name of the credentials associated with the enterprise.
                    This field is now deprecated. Use CredentialsID instead. This field will be
                    removed in v0.2.0.
                type: string
                x-go-name: CredentialsName
            endpoint:
                $ref: '#/definitions/ForgeEndpoint'
            events:
                items:
                    $ref: '#/definitions/EntityEvent'
                type: array
                x-go-name: Events
            id:
                type: string
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            owner:
                type: string
                x-go-name: Owner
            pool:
                items:
                    $ref: '#/definitions/Pool'
                type: array
                x-go-name: Pools
            pool_balancing_type:
                $ref: '#/definitions/PoolBalancerType'
            pool_manager_status:
                $ref: '#/definitions/PoolManagerStatus'
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
    Tag:
        properties:
            id:
                type: string
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/cloudbase/garm/params
...
info:
    description: The Garm API generated using go-swagger.
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0
    title: Garm API.
    version: 1.0.0
paths:
...
    /gitea/credentials:
        get:
            operationId: ListGiteaCredentials
            responses:
                "200":
                    description: Credentials
                    schema:
                        $ref: '#/definitions/Credentials'
                "400":
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: List all credentials.
            tags:
                - credentials
        post:
            operationId: CreateGiteaCredentials
            parameters:
                - description: Parameters used when creating a Gitea credential.
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/CreateGiteaCredentialsParams'
                    description: Parameters used when creating a Gitea credential.
                    type: object
            responses:
                "200":
                    description: ForgeCredentials
                    schema:
                        $ref: '#/definitions/ForgeCredentials'
                "400":
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: Create a Gitea credential.
            tags:
                - credentials
...
    /github/credentials:
        get:
            operationId: ListCredentials
            responses:
                "200":
                    description: Credentials
                    schema:
                        $ref: '#/definitions/Credentials'
                "400":
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: List all credentials.
            tags:
                - credentials
        post:
            operationId: CreateCredentials
            parameters:
                - description: Parameters used when creating a GitHub credential.
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/CreateGithubCredentialsParams'
                    description: Parameters used when creating a GitHub credential.
                    type: object
            responses:
                "200":
                    description: ForgeCredentials
                    schema:
                        $ref: '#/definitions/ForgeCredentials'
                "400":
                    description: APIErrorResponse
                    schema:
                        $ref: '#/definitions/APIErrorResponse'
            summary: Create a GitHub credential.
            tags:
                - credentials
...
produces:
    - application/json
security:
    - Bearer: []
securityDefinitions:
    Bearer:
        description: 'The token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"